# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Siklas API Documentation",
        "description": "API documentation for the Siklas project",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/riwayat-klasifikasi": {
            "get": {
                "tags": [
                    "Riwayat Klasifikasi"
                ],
                "summary": "List classification history",
                "description": "Get classification history based on the authenticated user role. Admin sees all, Kaprodi sees program-specific history, and others see their own.",
                "operationId": "listRiwayatKlasifikasi",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar riwayat klasifikasi berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Riwayat Klasifikasi"
                ],
                "summary": "Create a new classification record",
                "operationId": "storeRiwayatKlasifikasi",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "prediksi_topik"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "judul": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "abstrak": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "prediksi_topik": {
                                        "type": "string"
                                    },
                                    "confidence_score": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Riwayat klasifikasi berhasil disimpan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/riwayat-klasifikasi/{id}": {
            "get": {
                "tags": [
                    "Riwayat Klasifikasi"
                ],
                "summary": "Get a single classification record",
                "operationId": "getRiwayatKlasifikasi",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Riwayat klasifikasi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Riwayat Klasifikasi"
                ],
                "summary": "Delete a classification record",
                "operationId": "deleteRiwayatKlasifikasi",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Riwayat klasifikasi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}